feat: cleaner organization for code
