+++ cve_bin_tool/sbom_manager/__init__.py
@@ -16,7 +16,7 @@ from packageurl import PackageURL
 from cve_bin_tool.cvedb import CVEDB
 from cve_bin_tool.input_engine import TriageData
 from cve_bin_tool.log import LOGGER
-from cve_bin_tool.util import ProductInfo, Remarks
+from cve_bin_tool.util import ProductInfo, Remarks, find_product_location
 from cve_bin_tool.validator import validate_cyclonedx, validate_spdx
 
 from .swid_parser import SWIDParser
@@ -44,33 +44,6 @@ class SBOMManager:
 
         # Connect to the database
         self.cvedb = CVEDB(version_check=False)
-
-    def find_product_location(self, product_name):
-        for path in sys.path:
-            product_location = Path(path) / product_name
-            if product_location.exists():
-                return str(product_location)
-
-        known_installation_directories = [
-            "/usr/local/bin",
-            "/usr/local/sbin",
-            "/usr/bin",
-            "/opt",
-            "/usr/sbin",
-            "/usr/local/lib",
-            "/usr/lib",
-            "/usr/local/share",
-            "/usr/share",
-            "/usr/local/include",
-            "/usr/include",
-        ]
-
-        for directory in known_installation_directories:
-            product_location = Path(directory) / product_name
-            if product_location.exists():
-                return str(product_location)
-
-        return None
 
     def scan_file(self) -> dict[ProductInfo, TriageData]:
         self.logger.debug(
@@ -101,7 +74,7 @@ class SBOMManager:
                     # Now add vendor to create product record....
                     vendor_set = self.get_vendor(product)
                     for vendor in vendor_set:
-                        location = self.find_product_location(product)
+                        location = find_product_location(product)
                         if location is None:
                             location = "NotFound"
                         LOGGER.info(f"version_scan loca = {location}")
