+++ cve_bin_tool/util.py
@@ -276,6 +276,34 @@ def make_http_requests(attribute, **kwargs):
         LOGGER.error(ve)
 
 
+def find_product_location(self, product_name):
+    for path in sys.path:
+        product_location = Path(path) / product_name
+        if product_location.exists():
+            return str(product_location)
+
+    known_installation_directories = [
+        "/usr/local/bin",
+        "/usr/local/sbin",
+        "/usr/bin",
+        "/opt",
+        "/usr/sbin",
+        "/usr/local/lib",
+        "/usr/lib",
+        "/usr/local/share",
+        "/usr/share",
+        "/usr/local/include",
+        "/usr/include",
+    ]
+
+    for directory in known_installation_directories:
+        product_location = Path(directory) / product_name
+        if product_location.exists():
+            return str(product_location)
+
+    return None
+
+
 class DirWalk:
     """
     for filename in DirWalk('*.c').walk(roots):
