+++ cve_bin_tool/version_scanner.py
@@ -17,7 +17,13 @@ from cve_bin_tool.file import is_binary
 from cve_bin_tool.log import LOGGER
 from cve_bin_tool.parsers.parse import parse, valid_files
 from cve_bin_tool.strings import parse_strings
-from cve_bin_tool.util import DirWalk, ProductInfo, ScanInfo, inpath
+from cve_bin_tool.util import (
+    DirWalk,
+    ProductInfo,
+    ScanInfo,
+    find_product_location,
+    inpath,
+)
 
 if sys.version_info >= (3, 10):
     from importlib import metadata as importlib_metadata
@@ -242,33 +248,6 @@ class VersionScanner:
 
         yield from self.run_checkers(filename, lines)
 
-    def find_product_location(self, product_name):
-        for path in sys.path:
-            product_location = Path(path) / product_name
-            if product_location.exists():
-                return str(product_location)
-
-        known_installation_directories = [
-            "/usr/local/bin",
-            "/usr/local/sbin",
-            "/usr/bin",
-            "/opt",
-            "/usr/sbin",
-            "/usr/local/lib",
-            "/usr/lib",
-            "/usr/local/share",
-            "/usr/share",
-            "/usr/local/include",
-            "/usr/include",
-        ]
-
-        for directory in known_installation_directories:
-            product_location = Path(directory) / product_name
-            if product_location.exists():
-                return str(product_location)
-
-        return None
-
     def run_checkers(self, filename: str, lines: str) -> Iterator[ScanInfo]:
         """process a Set of checker objects, run them on file lines,
         and yield information about detected products and versions.
@@ -304,7 +283,7 @@ class VersionScanner:
                             f'{file_path} {result["is_or_contains"]} {dummy_checker_name} {version}'
                         )
                         for vendor, product in checker.VENDOR_PRODUCT:
-                            location = self.find_product_location(product)
+                            location = find_product_location(product)
                             if location is None:
                                 location = file_path
                             yield ScanInfo(
